refactor(crawl): split engine/sitemap.rs into discover/backfill/filter + prune stale monolith allowlist (5.7.7)
sitemap.rs (721 lines) now holds only the shared bounded-HTTP fetch
helpers and re-exports; discovery, backfill, and scope/lastmod filtering
move to sitemap/{discover,backfill,filter}.rs. No behavior change —
public surface preserved via re-exports, all 114 engine tests pass.

Allowlist: drop 3 entries for deleted web files and 4 for Rust files now
under the 500-line cap or pattern-exempt (config/** glob). Remaining
entries: query.rs + runners/ingest.rs (unexpired, tracked) and the four
vendored lab-auth files (expire 2026-06-15).

Part of axon_rust beads follow-up to the expired-allowlist P1.
