+++ apps/web/openapi/axon.json
@@ -10,7 +10,7 @@
       "name": "MIT",
       "identifier": "MIT"
     },
-    "version": "5.7.6"
+    "version": "5.7.7"
   },
   "paths": {
     "/v1/ask": {
