+++ apps/web/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@axon/admin-panel",
-  "version": "5.7.6",
+  "version": "5.7.7",
   "private": true,
   "scripts": {
     "build": "next build",
